+++ app/src/main/java/org/simple/clinic/summary/bloodpressures/ui/BloodPressureSummaryViewInjector.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.summary.bloodpressures.view
+package org.simple.clinic.summary.bloodpressures.ui
 
 interface BloodPressureSummaryViewInjector {
   fun inject(target: BloodPressureSummaryView)